# client/docs/swagger-ui/swagger.yaml
swagger: '2.0'
info:
  title: Cosmos SDK Liquidity Module - REST and gRPC Gateway docs
  description: 'A REST interface for state queries, transactions'
  version: 1.2.0
paths:
...
  '/liquidity/pools/{pool_id}/batch/deposits':
    post:
      summary: Submit deposit to the liquidity pool batch
      operationId: DepositToLiquidityPoolApi
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              std_tx:
                type: object
                properties:
                  msg:
                    type: array
                    items:
                      type: string
                  fee:
                    type: object
                    properties:
                      gas:
                        type: string
                        format: uint64
                      amount:
                        type: array
                        items:
                          type: object
                          properties:
                            denom:
                              type: string
                            amount:
                              type: string
                          description: >-
                            Coin defines a coin with a denomination and an
                            amount.


                            NOTE: The amount field is an Int which implements
                            the custom method

                            signatures required by gogoproto.
                        title: amount is the amount of coins to be paid as a fee
                  memo:
                    type: string
                  signature:
                    type: object
                    properties:
                      signature:
                        type: string
                      pub_key:
                        type: object
                        properties:
                          type:
                            type: string
                          value:
                            type: string
                      account_number:
                        type: string
                        format: uint64
                      sequence:
                        type: string
                        format: uint64
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: pool_id
          in: path
          required: true
          type: string
          format: uint64
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              base_req:
                type: object
                properties:
                  from:
                    type: string
                  memo:
                    type: string
                  chain_id:
                    type: string
                  account_number:
                    type: string
                    format: uint64
                  sequence:
                    type: string
                    format: uint64
                  timeout_height:
                    type: string
                    format: uint64
                  fees:
                    type: array
                    items:
                      type: object
                      properties:
                        denom:
                          type: string
                        amount:
                          type: string
                      description: >-
                        Coin defines a coin with a denomination and an amount.


                        NOTE: The amount field is an Int which implements the
                        custom method

                        signatures required by gogoproto.
                  gas_prices:
                    type: array
                    items:
                      type: object
                      properties:
                        denom:
                          type: string
                        amount:
                          type: string
                      description: >-
                        DecCoin defines a coin with a denomination and a decimal
                        amount.


                        NOTE: The amount field is an Dec which implements the
                        custom method

                        signatures required by gogoproto.
                  gas:
                    type: string
                    format: uint64
                  gas_adjustment:
                    type: string
                  simulate:
                    type: boolean
              pool_id:
                type: string
                format: uint64
              msg:
                type: object
                properties:
                  depositor_address:
                    type: string
                  pool_id:
                    type: string
                    format: uint64
                  deposit_coins:
                    type: array
                    items:
                      type: object
                      properties:
                        denom:
                          type: string
                        amount:
                          type: string
                      description: >-
                        Coin defines a coin with a denomination and an amount.


                        NOTE: The amount field is an Int which implements the
                        custom method

                        signatures required by gogoproto.
      tags:
        - MsgApi
    get:
      summary: Get all pool batch deposit messages of the liquidity pool
      operationId: PoolBatchDepositMsgs
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              deposit_msgs:
                type: array
                items:
                  type: object
                  properties:
                    msg_height:
                      type: string
                      format: int64
                    msg_index:
                      type: string
                      format: uint64
                    executed:
                      type: boolean
                    succeed:
                      type: boolean
                    to_delete:
                      type: boolean
                    Msg:
                      type: object
                      properties:
                        depositor_address:
                          type: string
                        pool_id:
                          type: string
                          format: uint64
                        deposit_coins:
                          type: array
                          items:
                            type: object
                            properties:
                              denom:
                                type: string
                              amount:
                                type: string
                            description: >-
                              Coin defines a coin with a denomination and an
                              amount.


                              NOTE: The amount field is an Int which implements
                              the custom method

                              signatures required by gogoproto.
              pagination:
                description: pagination defines the pagination in the response.
                type: object
                properties:
                  next_key:
                    type: string
                    format: byte
                    title: |-
                      next_key is the key to be passed to PageRequest.key to
                      query the next page most efficiently
                  total:
                    type: string
                    format: uint64
                    title: >-
                      total is total number of results available if
                      PageRequest.count_total

                      was set, its value is undefined otherwise
                title: >-
                  PageResponse is to be embedded in gRPC response messages where
                  the corresponding

                  request message has used PageRequest
        default:
          description: An unexpected error response.
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: pool_id
          in: path
          required: true
          type: string
          format: uint64
        - name: pagination.key
          description: |-
            key is a value returned in PageResponse.next_key to begin
            querying the next page most efficiently. Only one of offset or key
            should be set.
          in: query
          required: false
          type: string
          format: byte
        - name: pagination.offset
          description: >-
            offset is a numeric offset that can be used when key is unavailable.

            It is less efficient than using key. Only one of offset or key
            should

            be set.
          in: query
          required: false
          type: string
          format: uint64
        - name: pagination.limit
          description: >-
            limit is the total number of results to be returned in the result
            page.

            If left empty it will default to a value to be set by each app.
          in: query
          required: false
          type: string
          format: uint64
        - name: pagination.count_total
          description: >-
            count_total is set to true  to indicate that the result set should
            include

            a count of the total number of items available for pagination in
            UIs. count_total

            is only respected when offset is used. It is ignored when key is
            set.
          in: query
          required: false
          type: boolean
      tags:
        - Query
...